refactor: Change foler model to type
