+++ app/(club)/club/members/util/mapRole.ts
@@ -1,4 +1,4 @@
-import { clubMemberRoleType } from "@/model/member";
+import { clubMemberRoleType } from "@/types/member";
 
 export const MAP_ROLE_TO_KO: Record<clubMemberRoleType, string> = {
   ADMIN: "관리자",
