+++ app/data/clubMembers.ts
@@ -1,4 +1,4 @@
-import { ClubMemberData } from "@/model/member";
+import { ClubMemberData } from "@/types/member";
 
 export const CLUB_MEMBER_DATA: ClubMemberData[] = [
   {
