+++ app/data/qna.ts
@@ -1,5 +1,5 @@
-import { ClubQuestionData } from "@/models/club";
-import { PageInfo } from "@/models/pageInfo";
+import { ClubQuestionData } from "@/types/club";
+import { PageInfo } from "@/types/pageInfo";
 
 export const pageInfo: PageInfo = {
   contentSize: 10,
